requirements.txt
numpy==1.21.6
pandas==1.3.5
scipy==1.7.3
numba==0.56.3
tqdm==4.64.1
ujson==5.5.0
python-igraph==0.9.11
matplotlib==3.5.3
scikit-learn==1.0.2
tensorboardX==2.5.1
texttable==1.6.4
torch==1.9.1+cu111
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
# faiss-gpu==1.7.2
einops==0.5.0
einops-exts==0.0.3
pickle5==0.0.11